refactor: Update tab painter method calls in tab labels and borders
