+++ src/main/java/krasa/editorGroups/tabs2/impl/KrEditorTabsBorder.kt
@@ -11,7 +11,6 @@ import krasa.editorGroups.tabs2.KrTabInfo
 import krasa.editorGroups.tabs2.KrTabsBorder
 import krasa.editorGroups.tabs2.KrTabsListener
 import krasa.editorGroups.tabs2.KrTabsPosition
-import krasa.editorGroups.tabs2.impl.painter.KrTabPainter
 import java.awt.*
 
 class KrEditorTabsBorder(tabs: KrTabsImpl) : KrTabsBorder(tabs) {
@@ -78,7 +77,7 @@ class KrEditorTabsBorder(tabs: KrTabsImpl) : KrTabsBorder(tabs) {
         color = JBUI.CurrentTheme.EditorTabs.borderColor()
       )
     } else {
-      KrTabPainter.paintBorderLine(
+      tabs.tabPainter.paintBorderLine(
         g = g,
         thickness = thickness,
         from = Point(x, y),
@@ -100,12 +99,12 @@ class KrEditorTabsBorder(tabs: KrTabsImpl) : KrTabsBorder(tabs) {
 
         for (eachRow in startRow until lastRow) {
           val yl = eachRow * tabs.headerFitSize!!.height + startY
-          KrTabPainter.paintBorderLine(g, thickness, Point(x, yl), Point(x + width, yl))
+          tabs.tabPainter.paintBorderLine(g, thickness, Point(x, yl), Point(x + width, yl))
         }
 
         if (!NewUI.isEnabled() || (tabs as? KrEditorTabs)?.shouldPaintBottomBorder() == true) {
           val yl = lastRow * tabs.headerFitSize!!.height + startY
-          KrTabPainter.paintBorderLine(g, thickness, Point(x, yl), Point(x + width, yl))
+          tabs.tabPainter.paintBorderLine(g, thickness, Point(x, yl), Point(x + width, yl))
         }
       }
 
@@ -113,24 +112,24 @@ class KrEditorTabsBorder(tabs: KrTabsImpl) : KrTabsBorder(tabs) {
         val rowCount = tabs.lastLayoutPass!!.rowCount
         for (rowInd in 0 until rowCount) {
           val curY = height - (rowInd + 1) * tabs.headerFitSize!!.height
-          KrTabPainter.paintBorderLine(g, thickness, Point(x, curY), Point(x + width, curY))
+          tabs.tabPainter.paintBorderLine(g, thickness, Point(x, curY), Point(x + width, curY))
         }
       }
 
       KrTabsPosition.right -> {
         val lx = firstLabel.x
-        KrTabPainter.paintBorderLine(g, thickness, Point(lx, y), Point(lx, y + height))
+        tabs.tabPainter.paintBorderLine(g, thickness, Point(lx, y), Point(lx, y + height))
       }
 
       KrTabsPosition.left -> {
         val bounds = firstLabel.bounds
         val i = bounds.x + bounds.width - thickness
-        KrTabPainter.paintBorderLine(g, thickness, Point(i, y), Point(i, y + height))
+        tabs.tabPainter.paintBorderLine(g, thickness, Point(i, y), Point(i, y + height))
       }
     }
 
     if (hasAnimation()) {
-      KrTabPainter.paintUnderline(
+      tabs.tabPainter.paintUnderline(
         tabs.position,
         calcRectangle() ?: return,
         thickness,
@@ -139,7 +138,7 @@ class KrEditorTabsBorder(tabs: KrTabsImpl) : KrTabsBorder(tabs) {
       )
     } else {
       val selectedLabel = tabs.selectedLabel ?: return
-      KrTabPainter.paintUnderline(
+      tabs.tabPainter.paintUnderline(
         tabs.position,
         selectedLabel.bounds,
         thickness,
